extracts PredictionPercentageInput into its own component
